CosyVoice/async_cosyvoice/runtime/async_grpc/requirements.txt
fastapi==0.112.0
uvicorn==0.24.0
pydantic==2.10.6
grpcio==1.57.0
grpcio-tools==1.57.0
numpy==1.24.3
soundfile==0.12.1